refactor: keep compute method as abstract instead
